[test] HttpStaticHandler 테스트 수정
HttpStaticHandler 와 HttpRequest 가 수정됨에 따라 테스트 코드도 수정하였습니다.
